This is getting annoying

+++ src/host/dmr/Control.h
@@ -137,7 +137,7 @@ namespace dmr
         uint32_t getLastSrcId(uint32_t slotNo) const;
 
         /// <summary>Helper to get the last received RSSI.</summary>
-        int32_t getLastRssi(uint32_t slotNo) const;
+        uint8_t getLastRssi(uint32_t slotNo) const;
 
     private:
         friend class Slot;
